Add files via upload
This Word document contains 241 plot points that constitute a complete story and provide options for different paths the story can take. They are organized into different categories, which correspond the folders they will be nested in in the program. You can also print them out and organize them physically. These plot points are labeled as belonging to "THE HERO."